Ensure that the padding's statuslines are empty.
I modify buffer's statuslines via autocmds to indicate their state via my StatusLineHighlight plugin (vimscript #3383). Together with :DistractFree, that leaves a visible statusline in the padding window above the main window.
To avoid any such interference from autocmds, move the modification of the &statusline option in s:SaveRestore() to before the creation of the padding windows, and explicitly set the local statusline for those (just to be explicit and sure), too.